Bug: If switching between two series, the mask list is corrupted.
If switiching between series, a mask is automaticly added in the viewerwidget. The loop, for checking the existance of the mask, starts not as expected at the first element
but somewhere in the list. This results in a discrepancy between the actual mask list in memory and the list in viewer.

Fix: Changed searchpoint of list.

Changes:
* libviewer/src/libviewer.c
  - List search point
  - added security to only display related (group_id) series
  - added security to only scroll in related (group_id) series

source/gui/mainwindow.c
  - Application setted tree as active mask. This was wrong it should be the serie
  - Cleaned up the source in case if someone clicked on the same serie as already was selected.